incusd/network/openvswitch: Rename to ovs
Sponsored-by: Luizalabs (https://luizalabs.com)
